Have to use callback_context to see which shapes changed
This is because, even if shapes were deleted by setting the layout.shapes field
of the returned figure to [] in the last callback, the graph.relayoutData
parameter still will contain shapes. So we check to see if the shapes changed
using the callback_context and if not we just pass the shapes stored before.