Use JSON.stringify to compare objects
This is because it seems that 2 objects with all the same fields and elements
are only equal if they share the same memory. So we convert to JSON-formatted
strings before comparing. This is OK in our case, because we're using JSON
stringifying and parsing to make copies of the objects anyways.